refactor(image): separate grid and preview display helpers
